+++ LICENSE.txt
@@ -1,6 +1,6 @@
 The MIT License (MIT)
 
-Copyright (c) 2019 Ole Friis Østergaard
+Copyright (c) 2019-2022 Ole Friis Østergaard and Tina Wuest
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
