--- lib/mos6502/th8.rb
@@ -1,5 +1,0 @@
-require 'th8/bus'
-require 'th8/cpu'
-
-module TH8
-end
