+++ lib/sidtool/mos6502/cpu.rb
@@ -1,0 +1,5 @@
+module Mos6502
+  class Cpu
+    
+  end
+end
