+++ lib/sidtool/mos6502/mos6502.rb
@@ -1,0 +1,3 @@
+require_relative 'cpu'
+require_relative 'cpudsl'
+require_relative 'cpu6502'
