+++ lib/sidtool/mos6502/th8.rb
@@ -1,0 +1,7 @@
+require_relative 'th8/bus'
+require_relative 'th8/cpu'
+
+module Mos6502
+  module TH8
+  end
+end
